changed admin dashboard and date format in match model

+++ src/app/admin/admin-manage-recharge/admin-manage-recharge.component.ts
@@ -4,6 +4,7 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
+import { ConfirmBoxComponent } from 'src/app/common/components/confirm-box/confirm-box.component';
 import { LoadingComponent } from 'src/app/common/components/loading/loading.component';
 import { SnackbarComponent } from 'src/app/common/components/snackbar/snackbar.component';
 import { getErrorMessage, NO_RESP } from 'src/app/common/constants/error-message';
@@ -21,7 +22,7 @@ export class AdminManageRechargeComponent implements OnInit {
 
   rechargeData: RechargeModel[] = [];
 
-  displayedColumns: string[] = ['rechargeId' , 'userId' , 'userName', 'points', 'rechargeDate', 'delete'];
+  displayedColumns: string[] = ['rechargeId', 'userId', 'userName', 'points', 'rechargeDate', 'delete'];
 
   dataSource: MatTableDataSource<RechargeModel> = new MatTableDataSource();
   // dataSource = new MatTableDataSource<RechargeModel>(this.rechargeData);
@@ -35,7 +36,7 @@ export class AdminManageRechargeComponent implements OnInit {
   ) { }
 
   async ngOnInit() {
-    
+
     this.rechargeData = await this.getRecharge();
     console.log(this.rechargeData);
 
@@ -44,42 +45,50 @@ export class AdminManageRechargeComponent implements OnInit {
     this.dataSource.paginator = this.paginator;
   }
 
-  applyFilter(event:Event) {
+  applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
-  async deleteRecharge(rechargeData:RechargeModel): Promise<any>
-  {
-    let panelClass = 'green';
-    let snackbarMsg = '';
-    let snackbarRef = null;
-    const dialogRef = this.dialog.open(LoadingComponent, { disableClose: true });
-    // let recharegeModel: RechargeModel[] = [];
-    let msg;
-    let resp = null;
-    try {
-      resp = await this.rechargeservice.deleteRecharge(rechargeData.rechargeId);
-      msg = resp.body.message;
-      if (msg) {
-        dialogRef.close();
-        snackbarMsg = msg;
-      location.reload();
-      } else {
-        snackbarMsg = NO_RESP;
-        panelClass = 'red';
+  async deleteRecharge(rechargeData: RechargeModel): Promise<any> {
+    if (this.dialog.openDialogs.length == 0) {
+      const dialogRef1 = this.dialog.open(ConfirmBoxComponent, {
+        panelClass: 'no-padding-dialog',
+        data: 'Think Twice Before Deleting'
+      });
+      const closeResp = await dialogRef1.afterClosed().toPromise();
+      if (closeResp) {
+        let panelClass = 'green';
+        let snackbarMsg = '';
+        let snackbarRef = null;
+        const dialogRef = this.dialog.open(LoadingComponent, { disableClose: true });
+        // let recharegeModel: RechargeModel[] = [];
+        let msg;
+        let resp = null;
+        try {
+          resp = await this.rechargeservice.deleteRecharge(rechargeData.rechargeId);
+          msg = resp.body.message;
+          if (msg) {
+            dialogRef.close();
+            snackbarMsg = msg;
+            location.reload();
+          } else {
+            snackbarMsg = NO_RESP;
+            panelClass = 'red';
+          }
+        } catch (ex) {
+          snackbarMsg = getErrorMessage(ex);
+          panelClass = 'red';
+        } finally {
+          dialogRef.close();
+        }
+        if (snackbarMsg) {
+          snackbarRef = this.snackbar.openFromComponent(SnackbarComponent,
+            getSnackbarProperties(snackbarMsg, panelClass));
+        }
+        return [];
       }
-    } catch (ex) {
-      snackbarMsg = getErrorMessage(ex);
-      panelClass = 'red';
-    } finally {
-      dialogRef.close();
     }
-    if (snackbarMsg) {
-      snackbarRef = this.snackbar.openFromComponent(SnackbarComponent,
-        getSnackbarProperties(snackbarMsg, panelClass));
-    }
-    return [];
   }
 
   async getRecharge(): Promise<any> {
