+++ src/app/admin/admin-manage-team/admin-manage-team.component.ts
@@ -4,6 +4,7 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
+import { ConfirmBoxComponent } from 'src/app/common/components/confirm-box/confirm-box.component';
 import { LoadingComponent } from 'src/app/common/components/loading/loading.component';
 import { SnackbarComponent } from 'src/app/common/components/snackbar/snackbar.component';
 import { getErrorMessage, NO_RESP } from 'src/app/common/constants/error-message';
@@ -44,42 +45,50 @@ export class AdminManageTeamComponent implements OnInit {
     this.dataSource.paginator = this.paginator;
   }
 
-  applyFilter(event:Event) {
+  applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
-  async deleteTeam(teamData:TeamModel): Promise<any>
-  {
-    let panelClass = 'green';
-    let snackbarMsg = '';
-    let snackbarRef = null;
-    const dialogRef = this.dialog.open(LoadingComponent, { disableClose: true });
-    // let recharegeModel: RechargeModel[] = [];
-    let msg;
-    let resp = null;
-    try {
-      resp = await this.teamservice.deleteTeam(teamData.teamId);
-      msg = resp.body.message;
-      if (msg) {
-        dialogRef.close();
-        snackbarMsg = msg;
-      location.reload();
-      } else {
-        snackbarMsg = NO_RESP;
-        panelClass = 'red';
+  async deleteTeam(teamData: TeamModel): Promise<any> {
+    if (this.dialog.openDialogs.length == 0) {
+      const dialogRef1 = this.dialog.open(ConfirmBoxComponent, {
+        panelClass: 'no-padding-dialog',
+        data: 'Think Twice Before Deleting'
+      });
+      const closeResp = await dialogRef1.afterClosed().toPromise();
+      if (closeResp) {
+        let panelClass = 'green';
+        let snackbarMsg = '';
+        let snackbarRef = null;
+        const dialogRef = this.dialog.open(LoadingComponent, { disableClose: true });
+        // let recharegeModel: RechargeModel[] = [];
+        let msg;
+        let resp = null;
+        try {
+          resp = await this.teamservice.deleteTeam(teamData.teamId);
+          msg = resp.body.message;
+          if (msg) {
+            dialogRef.close();
+            snackbarMsg = msg;
+            location.reload();
+          } else {
+            snackbarMsg = NO_RESP;
+            panelClass = 'red';
+          }
+        } catch (ex) {
+          snackbarMsg = getErrorMessage(ex);
+          panelClass = 'red';
+        } finally {
+          dialogRef.close();
+        }
+        if (snackbarMsg) {
+          snackbarRef = this.snackbar.openFromComponent(SnackbarComponent,
+            getSnackbarProperties(snackbarMsg, panelClass));
+        }
+        return [];
       }
-    } catch (ex) {
-      snackbarMsg = getErrorMessage(ex);
-      panelClass = 'red';
-    } finally {
-      dialogRef.close();
     }
-    if (snackbarMsg) {
-      snackbarRef = this.snackbar.openFromComponent(SnackbarComponent,
-        getSnackbarProperties(snackbarMsg, panelClass));
-    }
-    return [];
   }
 
   async getTeams(): Promise<any> {
