+++ src/app/admin/admin-manage-user/admin-manage-user.component.ts
@@ -4,12 +4,14 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
+import { ConfirmBoxComponent } from 'src/app/common/components/confirm-box/confirm-box.component';
 import { LoadingComponent } from 'src/app/common/components/loading/loading.component';
 import { SnackbarComponent } from 'src/app/common/components/snackbar/snackbar.component';
 import { getErrorMessage, NO_RESP } from 'src/app/common/constants/error-message';
 import { getSnackbarProperties } from 'src/app/common/constants/snackbar-properties';
 import { UserModel } from 'src/app/common/model/user/user-model';
 import { UserService } from 'src/app/common/service/user_service/user.service';
+import { AssignRoleComponent } from '../CRUD/assign-role/assign-role.component';
 import { UserInsertComponent } from '../CRUD/user-insert/user-insert.component';
 
 
@@ -45,7 +47,7 @@ export class AdminManageUserComponent implements OnInit {
     this.dataSource.paginator = this.paginator;
   }
 
-  applyFilter(event:Event) {
+  applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
@@ -116,37 +118,45 @@ export class AdminManageUserComponent implements OnInit {
     return [];
   }
 
-  async deleteUser(userData:UserModel): Promise<any>
-  {
-    let panelClass = 'green';
-    let snackbarMsg = '';
-    let snackbarRef = null;
-    const dialogRef = this.dialog.open(LoadingComponent, { disableClose: true });
-    // let userModel: UserModel[] = [];
-    let msg;
-    let resp = null;
-    try {
-      resp = await this.userservice.deleteUser(userData.userId);
-      msg = resp.body.message;
-      if (msg) {
-        dialogRef.close();
-        snackbarMsg = msg;
-      location.reload();
-      } else {
-        snackbarMsg = NO_RESP;
-        panelClass = 'red';
+  async deleteUser(userData: UserModel): Promise<any> {
+    if (this.dialog.openDialogs.length == 0) {
+      const dialogRef1 = this.dialog.open(ConfirmBoxComponent, {
+        panelClass: 'no-padding-dialog',
+        data: 'Think Twice Before Deleting'
+      });
+      const closeResp = await dialogRef1.afterClosed().toPromise();
+      if (closeResp) {
+        let panelClass = 'green';
+        let snackbarMsg = '';
+        let snackbarRef = null;
+        const dialogRef = this.dialog.open(LoadingComponent, { disableClose: true });
+        // let userModel: UserModel[] = [];
+        let msg;
+        let resp = null;
+        try {
+          resp = await this.userservice.deleteUser(userData.userId);
+          msg = resp.body.message;
+          if (msg) {
+            snackbarMsg = msg;
+            dialogRef.close();
+            location.reload();
+          } else {
+            snackbarMsg = NO_RESP;
+            panelClass = 'red';
+          }
+        } catch (ex) {
+          snackbarMsg = getErrorMessage(ex);
+          panelClass = 'red';
+        } finally {
+          dialogRef.close();
+        }
+        if (snackbarMsg) {
+          snackbarRef = this.snackbar.openFromComponent(SnackbarComponent,
+            getSnackbarProperties(snackbarMsg, panelClass));
+        }
+        return [];
       }
-    } catch (ex) {
-      snackbarMsg = getErrorMessage(ex);
-      panelClass = 'red';
-    } finally {
-      dialogRef.close();
     }
-    if (snackbarMsg) {
-      snackbarRef = this.snackbar.openFromComponent(SnackbarComponent,
-        getSnackbarProperties(snackbarMsg, panelClass));
-    }
-    return [];
   }
 
 
@@ -163,6 +173,17 @@ export class AdminManageUserComponent implements OnInit {
   updateUserForm(data: any) {
     const dialogRef = this.dialog.open(UserInsertComponent,
       { panelClass: 'no-padding-dialog', disableClose: true, data });
+    dialogRef.afterClosed().toPromise().then(data => {
+      if (data) {
+        this.userData.push(data);
+        this.dataSource.data = this.userData;
+      }
+    });
+  }
+
+  openAssignForm() {
+    const dialogRef = this.dialog.open(AssignRoleComponent,
+      { panelClass: 'no-padding-dialog', disableClose: true });
     dialogRef.afterClosed().toPromise().then(data => {
       if (data) {
         this.userData.push(data);
